requirements.txt
discord.py
colorama~=0.4.4
Discord-Anti-Spam
captcha~=0.3